+++ examples/resources/elasticstack_kibana_osquery_saved_query/import.sh
@@ -1,0 +1,1 @@
+terraform import elasticstack_kibana_osquery_saved_query.list_processes <space id>/<saved query id>